Bms 77-78 clean
added handleUserSubmission.js
move stuff around
